moved gviz tests to tests module and inherits from django.test.TestCase instead of unittest
